Migrate Logo tests
