--- packages/lab/src/__tests__/logo/Logo.test.tsx
@@ -1,27 +1,0 @@
-import { render, screen } from "@testing-library/react";
-import { Logo } from "../../logo";
-
-const CustomLogoImage = ({ src }: any) => <div>{src}</div>;
-
-describe("GIVEN a logo", () => {
-  const appTitle = "Test App Title";
-  const src = "test-src.svg";
-
-  beforeEach(() => {
-    render(
-      <Logo
-        src={src}
-        appTitle={appTitle}
-        LogoImageComponent={CustomLogoImage}
-      />
-    );
-  });
-
-  test("logo image is rendered", () => {
-    expect(screen.getByText(src)).toBeInTheDocument();
-  });
-
-  test("appTitle is rendered", () => {
-    expect(screen.getByText(appTitle)).toBeInTheDocument();
-  });
-});
